refactor(cli): split 953-line exploit_k8s.py into kimera/cli/ package
Extract 20+ Click commands into domain-specific modules:
  __init__.py (126) — group, config loading, subcommand registration
  assess.py    (36) — security assessment
  exploit.py  (163) — exploit demos, vuln application
  validate.py  (53) — validate-control wrapper
  manage.py   (181) — revert, rollback, status, apply
  generate.py (191) — LLM remediation generation + DT MCP context
  secure.py   (135) — secure, verify
  enforce.py   (75) — Cilium enforcement management

exploit_k8s.py reduced to 26-line shim for backward compatibility.
All modules under 300 lines. 387 tests, zero lint errors.
