+++ kimera/cli/exploit.py
@@ -1,0 +1,163 @@
+# Copyright 2025 Dynatrace LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import traceback
+
+import click
+
+from ..container.core.logger import console
+from . import EXPLOITS
+
+
+@click.command()
+@click.argument("exploit_type", required=False)
+@click.option(
+    "--mode",
+    type=click.Choice(["info", "vulnerable", "demo"]),
+    default="demo",
+)
+@click.pass_context
+def exploit(ctx: click.Context, exploit_type: str | None, mode: str) -> None:
+    """Run security exploit demonstrations with modes."""
+    config = ctx.obj["config"]
+    logger = ctx.obj["logger"]
+    k8s = ctx.obj["k8s"]
+
+    if not exploit_type:
+        console.print("[bold]Available Exploits:[/bold]")
+        for name, cls in EXPLOITS.items():
+            console.print(f"  • {name}: {cls.name} ({cls.risk_level})")
+        return
+
+    if exploit_type not in EXPLOITS:
+        logger.error(f"Unknown exploit type: {exploit_type}")
+        return
+
+    service_name = config.exploit_mappings.get(exploit_type)
+    if not service_name:
+        logger.error(f"No service mapping found for exploit type: {exploit_type}")
+        return
+
+    exploit_class = EXPLOITS[exploit_type]
+    exploit_instance = exploit_class(k8s, service_name, logger)
+
+    try:
+        if mode == "info":
+            exploit_instance.show_info()
+        elif mode == "vulnerable":
+            exploit_instance.make_vulnerable()
+        elif mode == "demo":
+            exploit_instance.run_interactive()
+    except Exception as e:
+        logger.error(f"Failed to run exploit: {e}")
+        if config.debug:
+            traceback.print_exc()
+
+
+@click.command()
+@click.pass_context
+def exploit_all(ctx: click.Context) -> None:
+    """Run all exploit demonstrations."""
+    config = ctx.obj["config"]
+    logger = ctx.obj["logger"]
+    k8s = ctx.obj["k8s"]
+
+    logger.info("Running all container security exploits...")
+
+    for exploit_type, service in config.exploit_mappings.items():
+        if exploit_type in EXPLOITS:
+            console.print(f"\n[bold]=== Exploit: {exploit_type} ===[/bold]")
+            exploit_class = EXPLOITS[exploit_type]
+            exploit_instance = exploit_class(k8s, service, logger)
+            _ = exploit_instance.demonstrate()
+
+            if not click.confirm("\nContinue to next exploit?", default=True):
+                break
+
+    logger.success("All exploits demonstrated")
+
+
+@click.command()
+@click.pass_context
+def vuln(ctx: click.Context) -> None:
+    """Apply all standard vulnerabilities (DANGEROUS!)."""
+    config = ctx.obj["config"]
+    logger = ctx.obj["logger"]
+    k8s = ctx.obj["k8s"]
+
+    logger.warning("Applying all security vulnerabilities for demonstration...")
+    logger.warning("This will make services EXTREMELY VULNERABLE!")
+
+    if not config.dry_run:
+        response = click.confirm("Are you absolutely sure?", default=False)
+        if not response:
+            logger.info("Cancelled")
+            return
+
+    for et, service in config.exploit_mappings.items():
+        if et in EXPLOITS:
+            logger.warning(f"Applying {et} vulnerability to {service}...")
+            exploit_class = EXPLOITS[et]
+            exploit_obj = exploit_class(k8s, service, logger)
+            exploit_obj.make_vulnerable(dry_run=config.dry_run)
+
+    if not config.dry_run:
+        logger.warning("All vulnerabilities applied!")
+        console.print("\n⚠️  Services are now HIGHLY VULNERABLE:")
+        for et, service in config.exploit_mappings.items():
+            console.print(f"  • {service}: {et}")
+
+
+@click.command()
+@click.argument("service")
+@click.argument(
+    "vuln_type",
+    type=click.Choice(
+        ["privileged", "capabilities", "host-namespace", "no-limits", "no-network-policies", "rbac", "all"]
+    ),
+)
+@click.pass_context
+def vuln_service(ctx: click.Context, service: str, vuln_type: str) -> None:
+    """Apply vulnerability to a specific service."""
+    logger = ctx.obj["logger"]
+    k8s = ctx.obj["k8s"]
+
+    logger.info(f"Applying {vuln_type} vulnerability to {service}...")
+
+    vuln_alias_map: dict[str, str] = {
+        "privileged": "privileged-containers",
+        "capabilities": "dangerous-capabilities",
+        "host-namespace": "host-namespace-sharing",
+        "no-limits": "missing-resource-limits",
+        "no-network-policies": "missing-network-policies",
+        "rbac": "rbac-abuse",
+    }
+
+    try:
+        if vuln_type == "all":
+            for registry_key in vuln_alias_map.values():
+                exploit_class = EXPLOITS[registry_key]
+                exploit_obj = exploit_class(k8s, service, logger)
+                exploit_obj.make_vulnerable(dry_run=ctx.obj["config"].dry_run)
+        else:
+            registry_key = vuln_alias_map[vuln_type]
+            exploit_class = EXPLOITS[registry_key]
+            exploit_obj = exploit_class(k8s, service, logger)
+            exploit_obj.make_vulnerable(dry_run=ctx.obj["config"].dry_run)
+
+        logger.success(f"Vulnerability {vuln_type} applied to {service}.")
+    except Exception as e:
+        logger.error(f"Failed to apply vulnerability: {e}")
+        if ctx.obj["config"].debug:
+            traceback.print_exc()
