+++ kimera/exploit_k8s.py
@@ -12,939 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import sys
-import traceback
-from typing import Any
+# Shim for backward compatibility — all CLI logic lives in kimera/cli/
+from .cli import cli
 
-import click
-from dotenv import load_dotenv
-from rich.table import Table
 
-from .application.config.loader import ConfigLoader
-from .application.config.registry import ExploitRegistry
-from .application.config.schemas import ToolkitConfig
-from .container.assessment.scanner import SecurityScanner
-from .container.core.journal import clear_all, pending_operations, record_operation
-from .container.core.k8s_client import K8sClient
-from .container.core.logger import SecurityLogger, console, setup_logger
-from .container.infrastructure.enforcement import PolicyEnforcementManager
-from .container.infrastructure.resource_applier import ResourceApplier
-from .container.remediations.generator import LLMRemediationGenerator
-from .container.validation.engine import validate_controls
-
-# Load exploit registry from config/exploits/registry.yaml
-REGISTRY = ExploitRegistry()
-EXPLOITS = REGISTRY.classes
-
-
-def _load_config(
-    namespace: str,
-    profile: str | None,
-    debug: bool,
-    dry_run: bool,
-    verbose: bool,
-) -> ToolkitConfig:
-    """Load configuration with profile auto-detection and CLI overrides."""
-    overrides: dict[str, Any] = {}
-
-    if debug:
-        overrides["debug"] = True
-    if dry_run:
-        overrides["dry_run"] = True
-    if verbose:
-        overrides["verbose"] = True
-
-    # Auto-detect unguard profile when namespace is "unguard"
-    effective_profile = profile
-    if not effective_profile and namespace == "unguard":
-        effective_profile = "unguard"
-
-    if namespace != "default":
-        overrides["kubernetes"] = {"namespace": namespace}
-
-    loader = ConfigLoader()
-    return loader.load(profile=effective_profile, overrides=overrides)
-
-
-def _resolve_services(config: ToolkitConfig, k8s: K8sClient) -> list[str]:
-    """Return configured services or auto-discover from namespace."""
-    if config.services:
-        return config.services
-    try:
-        deployments = k8s.apps_v1.list_namespaced_deployment(config.namespace)
-        return [d.metadata.name for d in deployments.items]
-    except Exception:
-        return []
-
-
-@click.group()
-@click.option("--namespace", "-n", default="default", help="Target namespace")
-@click.option("--profile", "-p", default=None, help="Config profile (e.g., unguard)")
-@click.option("--debug", is_flag=True, help="Enable debug output")
-@click.option("--dry-run", is_flag=True, help="Preview changes without applying")
-@click.option("--verbose", "-v", is_flag=True, help="Enable verbose output")
-@click.pass_context
-def cli(
-    ctx: click.Context,
-    namespace: str,
-    profile: str | None,
-    debug: bool,
-    dry_run: bool,
-    verbose: bool,
-) -> None:
-    """Kubernetes Container Security Lab - Assessment and Exploitation Toolkit."""
-    ctx.ensure_object(dict)
-
-    config = _load_config(namespace, profile, debug, dry_run, verbose)
-    logger = SecurityLogger(setup_logger("k8s_security", debug=config.debug))
-
-    try:
-        k8s_client = K8sClient(namespace=config.namespace, logger=logger, verbose=config.verbose)
-    except Exception as e:
-        logger.error(f"Failed to initialize Kubernetes client: {e}")
-        sys.exit(1)
-
-    ctx.obj["config"] = config
-    ctx.obj["logger"] = logger
-    ctx.obj["k8s"] = k8s_client
-
-    # Banner for main commands
-    if ctx.invoked_subcommand in ["assess", "exploit", "secure", "verify"]:
-        from .banner import show_banner
-
-        show_banner(config.namespace)
-
-
-@cli.command()
-@click.argument("service", required=False)
-@click.pass_context
-def assess(ctx: click.Context, service: str | None) -> None:
-    """Assess security posture of services."""
-    config = ctx.obj["config"]
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    scanner = SecurityScanner(k8s, logger)
-
-    if service:
-        scanner.assess_service(service)
-    else:
-        services = _resolve_services(config, k8s)
-        scanner.assess_all_services(services)
-
-
-@cli.command()
-@click.pass_context
-def vuln(ctx: click.Context) -> None:
-    """Apply all standard vulnerabilities (DANGEROUS!)."""
-    config = ctx.obj["config"]
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    logger.warning("Applying all security vulnerabilities for demonstration...")
-    logger.warning("This will make services EXTREMELY VULNERABLE!")
-
-    if not config.dry_run:
-        response = click.confirm("Are you absolutely sure?", default=False)
-        if not response:
-            logger.info("Cancelled")
-            return
-
-    # Apply vulnerabilities
-    for exploit_type, service in config.exploit_mappings.items():
-        if exploit_type in EXPLOITS:
-            logger.warning(f"Applying {exploit_type} vulnerability to {service}...")
-            exploit_class = EXPLOITS[exploit_type]
-            exploit = exploit_class(k8s, service, logger)
-            exploit.make_vulnerable(dry_run=config.dry_run)
-
-    if not config.dry_run:
-        logger.warning("All vulnerabilities applied!")
-        console.print("\n⚠️  Services are now HIGHLY VULNERABLE:")
-        for exploit_type, service in config.exploit_mappings.items():
-            console.print(f"  • {service}: {exploit_type}")
-
-
-@cli.command()
-@click.argument("service")
-@click.argument(
-    "vuln_type",
-    type=click.Choice(
-        ["privileged", "capabilities", "host-namespace", "no-limits", "no-network-policies", "rbac", "all"]
-    ),
-)
-@click.pass_context
-def vuln_service(ctx: click.Context, service: str, vuln_type: str) -> None:
-    """Apply vulnerability to a specific service."""
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    logger.info(f"Applying {vuln_type} vulnerability to {service}...")
-
-    # Short aliases for CLI convenience → registry keys
-    vuln_alias_map: dict[str, str] = {
-        "privileged": "privileged-containers",
-        "capabilities": "dangerous-capabilities",
-        "host-namespace": "host-namespace-sharing",
-        "no-limits": "missing-resource-limits",
-        "no-network-policies": "missing-network-policies",
-        "rbac": "rbac-abuse",
-    }
-
-    try:
-        if vuln_type == "all":
-            for registry_key in vuln_alias_map.values():
-                exploit_class = EXPLOITS[registry_key]
-                exploit = exploit_class(k8s, service, logger)
-                exploit.make_vulnerable(dry_run=ctx.obj["config"].dry_run)
-        else:
-            registry_key = vuln_alias_map[vuln_type]
-            exploit_class = EXPLOITS[registry_key]
-            exploit = exploit_class(k8s, service, logger)
-            exploit.make_vulnerable(dry_run=ctx.obj["config"].dry_run)
-
-        logger.success(f"Vulnerability {vuln_type} applied to {service}.")
-    except Exception as e:
-        logger.error(f"Failed to apply vulnerability: {e}")
-        if ctx.obj["config"].debug:
-            traceback.print_exc()
-
-
-@cli.command()
-@click.argument("exploit_type", required=False)
-@click.option(
-    "--mode",
-    type=click.Choice(["info", "vulnerable", "demo"]),
-    default="demo",
-)
-@click.pass_context
-def exploit(ctx: click.Context, exploit_type: str | None, mode: str) -> None:
-    """Run security exploit demonstrations with modes."""
-    config = ctx.obj["config"]
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    if not exploit_type:
-        console.print("[bold]Available Exploits:[/bold]")
-        for name, cls in EXPLOITS.items():
-            console.print(f"  • {name}: {cls.name} ({cls.risk_level})")
-        return
-
-    if exploit_type not in EXPLOITS:
-        logger.error(f"Unknown exploit type: {exploit_type}")
-        return
-
-    service_name = config.exploit_mappings.get(exploit_type)
-    if not service_name:
-        logger.error(f"No service mapping found for exploit type: {exploit_type}")
-        return
-
-    exploit_class = EXPLOITS[exploit_type]
-    exploit_instance = exploit_class(k8s, service_name, logger)
-
-    try:
-        if mode == "info":
-            exploit_instance.show_info()
-        elif mode == "vulnerable":
-            exploit_instance.make_vulnerable()
-        elif mode == "demo":
-            exploit_instance.run_interactive()
-    except Exception as e:
-        logger.error(f"Failed to run exploit: {e}")
-        if config.debug:
-            traceback.print_exc()
-
-
-@cli.command()
-@click.pass_context
-def exploit_all(ctx: click.Context) -> None:
-    """Run all exploit demonstrations."""
-    config = ctx.obj["config"]
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    logger.info("Running all container security exploits...")
-
-    for exploit_type, service in config.exploit_mappings.items():
-        if exploit_type in EXPLOITS:
-            console.print(f"\n[bold]=== Exploit: {exploit_type} ===[/bold]")
-            exploit_class = EXPLOITS[exploit_type]
-            exploit_instance = exploit_class(k8s, service, logger)
-
-            _ = exploit_instance.demonstrate()
-
-            if not click.confirm("\nContinue to next exploit?", default=True):
-                break
-
-    logger.success("All exploits demonstrated")
-
-
-@cli.command()
-@click.argument("exploit_type", required=False)
-@click.option("--service", "-s", help="Override target service")
-@click.pass_context
-def secure(ctx: click.Context, exploit_type: str | None, service: str | None) -> None:
-    """Print remediation guidance for security issues.
-
-    Remediations are generated dynamically via 'kimera generate' and applied
-    via 'kimera apply'. Use 'kimera revert' to restore original state.
-    """
-    config = ctx.obj["config"]
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    if exploit_type:
-        if exploit_type not in EXPLOITS:
-            logger.error(f"Unknown exploit type: {exploit_type}")
-            return
-
-        target_service = service or config.exploit_mappings.get(exploit_type)
-        exploit_class = EXPLOITS[exploit_type]
-        exploit_instance = exploit_class(k8s, target_service, logger)
-        exploit_instance.make_secure(dry_run=config.dry_run)
-    else:
-        for et, svc in config.exploit_mappings.items():
-            if et in EXPLOITS:
-                exploit_class = EXPLOITS[et]
-                exploit_instance = exploit_class(k8s, svc, logger)
-                exploit_instance.make_secure(dry_run=config.dry_run)
-
-
-@cli.command()
-@click.pass_context
-def verify(ctx: click.Context) -> None:
-    """Verify security improvements."""
-    config = ctx.obj["config"]
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    logger.info("Verifying security improvements...")
-
-    all_secure = True
-    results = []
-    services = _resolve_services(config, k8s)
-
-    for service in services:
-        pod_name = k8s.find_pod_for_service(service)
-        if not pod_name:
-            results.append((service, "No pod found", False))
-            all_secure = False
-            continue
-
-        try:
-            pod = k8s.v1.read_namespaced_pod(pod_name, k8s.namespace)
-            issues = []
-
-            # Check privileged
-            container = pod.spec.containers[0]
-            if container.security_context and container.security_context.privileged:
-                issues.append("Privileged")
-
-            # Check capabilities
-            if (
-                container.security_context
-                and container.security_context.capabilities
-                and container.security_context.capabilities.add
-            ):
-                issues.append("Dangerous capabilities")
-
-            # Check host access
-            if pod.spec.host_pid or pod.spec.host_network or pod.spec.host_ipc:
-                issues.append("Host namespace access")
-
-            # Check resource limits
-            if not container.resources or not container.resources.limits:
-                issues.append("No resource limits")
-
-            if issues:
-                results.append((service, ", ".join(issues), False))
-                all_secure = False
-            else:
-                results.append((service, "Secure", True))
-
-        except Exception as e:
-            results.append((service, f"Error: {e}", False))
-            all_secure = False
-
-    table = Table(title="Security Verification Results")
-    table.add_column("Service", style="cyan")
-    table.add_column("Status", style="white")
-    table.add_column("Secure", style="white")
-
-    for service, status, secure in results:
-        table.add_row(service, status, "✅" if secure else "❌")
-
-    console.print(table)
-
-    # Check network policies
-    policies = k8s.list_network_policies(k8s.namespace)
-    if policies:
-        console.print(f"\n✅ {len(policies)} NetworkPolicies in namespace {k8s.namespace}")
-    else:
-        console.print(f"\n❌ No NetworkPolicies in namespace {k8s.namespace}")
-        all_secure = False
-
-    if all_secure:
-        logger.secure("All security verifications passed! 🎉")
-        console.print("\nThe containers are now properly secured with:")
-        console.print("  • Non-privileged execution")
-        console.print("  • Dropped capabilities")
-        console.print("  • Resource limits")
-        console.print("  • Host isolation")
-        console.print("  • Disabled privilege escalation")
-        console.print("  • Network segmentation via NetworkPolicies")
-    else:
-        logger.warning("Some security issues remain. Please review the output above.")
-
-
-@cli.command()
-@click.argument("service", required=False)
-@click.option("--revision", "-r", type=int, help="Rollback to specific revision")
-@click.pass_context
-def rollback(ctx: click.Context, service: str | None, revision: int | None) -> None:
-    """Rollback deployment to a previous revision.
-
-    For a higher-level undo that handles both deployments and network policies,
-    use 'kimera revert' instead.
-    """
-    config = ctx.obj["config"]
-    k8s = ctx.obj["k8s"]
-
-    if service:
-        k8s.rollback_deployment(service, revision)
-    else:
-        services = _resolve_services(config, k8s)
-        for svc in services:
-            k8s.rollback_deployment(svc, revision)
-
-
-@cli.command()
-@click.pass_context
-def rollback_original(ctx: click.Context) -> None:
-    """Rollback all services to original version (revision 1)."""
-    config = ctx.obj["config"]
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    logger.info("Rolling back all services to original versions...")
-    services = _resolve_services(config, k8s)
-
-    for service in services:
-        k8s.rollback_deployment(service, revision=1)
-
-
-@cli.command()
-@click.argument("exploit_type", required=False)
-@click.pass_context
-def revert(ctx: click.Context, exploit_type: str | None) -> None:
-    """Revert changes made by kimera, restoring original deployment state.
-
-    Without arguments, reverts ALL recorded operations in the current namespace.
-    With an exploit type, reverts only that exploit.
-
-    Unlike 'secure' (which applies a hardened config), 'revert' rolls back
-    deployments to their previous revision and removes toolkit-managed
-    NetworkPolicies.
-    """
-    config = ctx.obj["config"]
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    if exploit_type:
-        # Revert a specific exploit type
-        if exploit_type not in EXPLOITS:
-            logger.error(f"Unknown exploit type: {exploit_type}")
-            return
-
-        service = config.exploit_mappings.get(exploit_type)
-        if not service:
-            logger.error(f"No service mapping found for {exploit_type}")
-            return
-
-        exploit_class = EXPLOITS[exploit_type]
-        exploit_instance = exploit_class(k8s, service, logger)
-        exploit_instance.revert(dry_run=config.dry_run)
-        return
-
-    # Revert all recorded operations
-    ops = pending_operations(namespace=config.namespace)
-    if not ops:
-        logger.info("No recorded operations to revert")
-        return
-
-    logger.info(f"Reverting {len(ops)} recorded operation(s)...")
-
-    # Deduplicate by (exploit_type, service) — revert each pair once
-    seen: set[tuple[str, str]] = set()
-    for op in ops:
-        key = (op["exploit_type"], op["service"])
-        if key in seen:
-            continue
-        seen.add(key)
-
-        action = op.get("action", "make_vulnerable")
-        et = op["exploit_type"]
-        svc = op["service"]
-
-        if action == "apply":
-            # Applied resources — delete all toolkit-labeled resources in namespace
-            _revert_applied_resources(k8s, logger, config.namespace, config.dry_run)
-        elif et in REGISTRY:
-            entry = REGISTRY[et]
-            exploit_instance = entry.cls(k8s, svc, logger)
-            exploit_instance.revert(dry_run=config.dry_run)
-        else:
-            logger.warning(f"Unknown exploit type in journal: {et}")
-
-    if not config.dry_run:
-        clear_all()
-        logger.success("All operations reverted")
-
-
-@cli.command()
-@click.pass_context
-def status(ctx: click.Context) -> None:
-    """Show current deployment and security status."""
-    config = ctx.obj["config"]
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    logger.info("Checking current status...")
-
-    console.print("\n[bold]Namespace Status[/bold]")
-    try:
-        ns = k8s.v1.read_namespace(config.namespace)
-        console.print(f"✅ Namespace '{ns}' exists")
-    except Exception:
-        console.print("❌ Namespace not found")
-        return
-
-    console.print("\n[bold]Deployment Status[/bold]")
-    deployments = k8s.apps_v1.list_namespaced_deployment(config.namespace)
-
-    for dep in deployments.items:
-        ready = f"{dep.status.ready_replicas or 0}/{dep.spec.replicas}"
-        if dep.status.ready_replicas == dep.spec.replicas:
-            console.print(f"✅ {dep.metadata.name}: {ready}")
-        else:
-            console.print(f"⚠️  {dep.metadata.name}: {ready}")
-
-    console.print("\n[bold]Security Status[/bold]")
-    scanner = SecurityScanner(k8s, logger)
-    services = _resolve_services(config, k8s)
-    scanner.quick_security_check(services)
-
-
-@cli.command()
-@click.argument("service")
-@click.pass_context
-def secure_service(ctx: click.Context, service: str) -> None:
-    """Print remediation guidance for a specific service."""
-    config = ctx.obj["config"]
-    ns = config.namespace
-    console.print(f"\n[bold]Remediation for {service}[/bold]\n")
-    console.print(f"  kimera -n {ns} generate --type all --apply\n")
-    console.print("[dim]Or target a specific exploit type:[/dim]")
-    console.print(f"  kimera -n {ns} generate --type privileged-containers --apply")
-    console.print(f"  kimera -n {ns} generate --type dangerous-capabilities --apply")
-    console.print(f"  kimera -n {ns} generate --type host-namespace-sharing --apply")
-    console.print(f"  kimera -n {ns} generate --type missing-resource-limits --apply")
-    console.print(f"  kimera -n {ns} generate --type missing-network-policies --apply\n")
-
-
-@cli.command("apply")
-@click.argument("file", type=click.Path(exists=True, dir_okay=False))
-@click.pass_context
-def apply_resources(ctx: click.Context, file: str) -> None:
-    """Apply Kubernetes resources from a YAML file.
-
-    Reads a multi-document YAML file, validates each resource, injects
-    the kimera managed-by label, and applies to the cluster. All applied
-    resources are tracked for clean revert via 'kimera revert'.
-    """
-    config = ctx.obj["config"]
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    applier = ResourceApplier(k8s, logger)
-    applied, _ = applier.apply_from_file(file, config.namespace, dry_run=config.dry_run)
-
-    if applied > 0 and not config.dry_run:
-        record_operation("apply", "applied-resources", str(file), config.namespace)
-
-
-@cli.command("validate-control")
-@click.option(
-    "--type",
-    "control_type",
-    default="all",
-    show_default=True,
-    type=click.Choice(["network-policy", "admission", "rbac", "all"]),
-    help="Type of security control to validate.",
-)
-@click.option(
-    "--json",
-    "output_json",
-    is_flag=True,
-    default=False,
-    help="Output results as JSON.",
-)
-@click.pass_context
-def validate_control(ctx: click.Context, control_type: str, output_json: bool) -> None:
-    r"""Validate that security controls actually block attacks.
-
-    Tests whether NetworkPolicies, admission controllers (PSA, Kyverno,
-    Gatekeeper), and RBAC restrictions hold by:
-
-    \b
-      network-policy: Deploy ephemeral probe pod, test connectivity
-      admission:      Submit test resources via --dry-run=server
-      rbac:           Check SA permissions via role/binding analysis
-
-    Safe for production: uses dry-run and ephemeral resources only.
-    """
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    validate_controls(k8s, logger, control_type=control_type, output_json=output_json)
-
-
-@cli.group()
-@click.pass_context
-def enforce(ctx: click.Context) -> None:
-    """Manage NetworkPolicy enforcement (Cilium)."""
-    pass
-
-
-@enforce.command("enable")
-@click.pass_context
-def enforce_enable(ctx: click.Context) -> None:
-    """Check for Cilium NetworkPolicy enforcement; print install guidance if not found."""
-    config = ctx.obj["config"]
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    manager = PolicyEnforcementManager(k8s, logger)
-    manager.enable(dry_run=config.dry_run)
-
-
-@enforce.command("disable")
-@click.pass_context
-def enforce_disable(ctx: click.Context) -> None:
-    """Show guidance for removing Cilium and disabling NetworkPolicy enforcement."""
-    config = ctx.obj["config"]
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    manager = PolicyEnforcementManager(k8s, logger)
-
-    if not manager.is_enabled():
-        logger.info("Cilium is not running — no enforcement to disable")
-        return
-
-    manager.disable(dry_run=config.dry_run)
-
-
-@enforce.command("status")
-@click.pass_context
-def enforce_status(ctx: click.Context) -> None:
-    """Show NetworkPolicy enforcement status."""
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    manager = PolicyEnforcementManager(k8s, logger)
-    status_info = manager.get_status()
-
-    if not status_info.get("installed"):
-        console.print("NetworkPolicy enforcement: [red]not installed[/red]")
-        console.print("\nRun 'kimera enforce enable' to check Cilium enforcement status.")
-        return
-
-    console.print("NetworkPolicy enforcement: [green]active[/green]")
-    console.print(f"  Image:     {status_info['image']}")
-    console.print(f"  Desired:   {status_info['desired']}")
-    console.print(f"  Ready:     {status_info['ready']}")
-    console.print(f"  Available: {status_info['available']}")
-
-
-@cli.command("generate")
-@click.option(
-    "--mode",
-    default="remediate",
-    show_default=True,
-    type=click.Choice(["remediate", "exploit"]),
-    help="Generation mode: 'remediate' fixes vulnerabilities, 'exploit' introduces them.",
-)
-@click.option(
-    "--type",
-    "exploit_type",
-    default="missing-network-policies",
-    show_default=True,
-    type=click.Choice(["all"] + REGISTRY.types),
-    help="Type of remediation or exploit to generate.",
-)
-@click.option(
-    "--service",
-    default=None,
-    help="Target service for exploit mode. If omitted, LLM recommends targets.",
-)
-@click.option(
-    "--output",
-    "-o",
-    default=None,
-    help="Output file path for generated YAML.",
-)
-@click.option(
-    "--model",
-    default="claude-sonnet-4-6",
-    show_default=True,
-    help="Anthropic model identifier.",
-)
-@click.option(
-    "--apply",
-    "apply_generated",
-    is_flag=True,
-    default=False,
-    help="Apply generated resources after generation.",
-)
-@click.option(
-    "--use-dt-mcp",
-    is_flag=True,
-    default=False,
-    help="Enrich context with Dynatrace MCP data (requires DT_ENVIRONMENT and DT_PLATFORM_TOKEN).",
-)
-@click.option(
-    "--dt-strategy",
-    default="targeted",
-    show_default=True,
-    type=click.Choice(["targeted", "llm-query", "davis"]),
-    help="DT data fetching strategy (requires --use-dt-mcp).",
-)
-@click.option(
-    "--yes",
-    "-y",
-    is_flag=True,
-    default=False,
-    help="Skip confirmation prompt when using --apply.",
-)
-@click.pass_context
-def generate(
-    ctx: click.Context,
-    mode: str,
-    exploit_type: str,
-    service: str | None,
-    output: str | None,
-    model: str,
-    apply_generated: bool,
-    use_dt_mcp: bool,
-    dt_strategy: str,
-    yes: bool,
-) -> None:
-    """Generate security remediations or exploit patches using an LLM.
-
-    Gathers Kubernetes context (workloads, security contexts, services),
-    optionally enriches with Dynatrace observability data via DT MCP,
-    and calls Anthropic Claude to generate YAML.
-
-    Modes:
-      remediate — generates hardened configs / network policies to fix vulnerabilities.
-      exploit   — generates JSON patches to introduce vulnerabilities for testing.
-
-    Requires the 'llm' extra: uv pip install 'kimera[llm]'
-    Set ANTHROPIC_API_KEY before running.
-    """
-    config = ctx.obj["config"]
-    logger = ctx.obj["logger"]
-    k8s = ctx.obj["k8s"]
-
-    # Default output file depends on mode
-    if output is None:
-        output = "kimera-exploit.yaml" if mode == "exploit" else "kimera-remediations.yaml"
-
-    # Validate --dt-strategy requires --use-dt-mcp
-    if dt_strategy != "targeted" and not use_dt_mcp:
-        logger.error("--dt-strategy requires --use-dt-mcp")
-        return
-
-    # Fetch DT MCP context if requested
-    smartscape_context: str | None = None
-    kspm_context: str | None = None
-
-    if use_dt_mcp:
-        dt_context = _fetch_dt_context(config, logger, dt_strategy, exploit_type, model)
-        if dt_context:
-            smartscape_context = dt_context.raw_smartscape or None
-            kspm_context = dt_context.raw_kspm or None
-
-    generator = LLMRemediationGenerator(
-        k8s=k8s,
-        logger=logger,
-        network_topology=config.network_topology,
-        model=model,
-    )
-
-    try:
-        if mode == "exploit":
-            yaml_output = generator.generate_exploit(
-                exploit_type=exploit_type,
-                service=service,
-                kspm_context=kspm_context,
-                smartscape_context=smartscape_context,
-            )
-        else:
-            yaml_output = generator.generate(
-                exploit_type=exploit_type,
-                kspm_context=kspm_context,
-                smartscape_context=smartscape_context,
-            )
-    except ImportError as e:
-        logger.error(str(e))
-        return
-    except ValueError as e:
-        logger.error(f"Generation failed: {e}")
-        return
-
-    # Write to file
-    from pathlib import Path  # noqa: PLC0415
-
-    Path(output).write_text(yaml_output)
-    label = "exploit patches" if mode == "exploit" else "remediations"
-    logger.success(f"Generated {label} written to {output}")
-    console.print(f"\n[dim]Preview:[/dim]\n{yaml_output[:2000]}")
-    if len(yaml_output) > 2000:
-        console.print(f"[dim]... ({len(yaml_output)} chars total, see {output})[/dim]")
-
-    if not apply_generated:
-        console.print(f"\n[dim]Re-run with --apply to apply, or use: kimera apply {output}[/dim]")
-        return
-
-    # Confirmation
-    if not yes and not config.dry_run:
-        if not click.confirm(f"\nApply {label} from {output}?", default=False):
-            logger.info("Cancelled")
-            return
-
-    applier = ResourceApplier(k8s, logger)
-
-    if mode == "exploit":
-        applied, _ = applier.apply_exploit_patches(output, config.namespace, dry_run=config.dry_run)
-        op_action = "make_vulnerable"
-    else:
-        applied, _ = applier.apply_from_file(output, config.namespace, dry_run=config.dry_run)
-        op_action = "apply"
-
-    if applied > 0 and not config.dry_run:
-        record_operation(op_action, exploit_type, output, config.namespace)
-
-
-def _fetch_dt_context(
-    config: ToolkitConfig,
-    logger: SecurityLogger,
-    strategy_name: str,
-    exploit_type: str,
-    model: str,
-) -> Any:
-    """Fetch Dynatrace context using the specified strategy.
-
-    Args:
-        config: Toolkit configuration.
-        logger: Security logger.
-        strategy_name: One of ``targeted``, ``llm-query``, ``davis``.
-        exploit_type: Exploit type to fetch context for.
-        model: Anthropic model identifier (for llm-query strategy).
-
-    Returns:
-        ``DtContext`` with fetched data, or ``None`` on failure.
-    """
-    import asyncio  # noqa: PLC0415
-    import os  # noqa: PLC0415
-
-    from .container.infrastructure.dt_data_models import DtContext  # noqa: PLC0415
-    from .container.infrastructure.dt_query_strategies import create_strategy  # noqa: PLC0415
-
-    dt_env = os.environ.get("DT_ENVIRONMENT", "")
-    dt_token = os.environ.get("DT_PLATFORM_TOKEN", "")
-
-    if not dt_env or not dt_token:
-        logger.warning(
-            "DT MCP requires DT_ENVIRONMENT and DT_PLATFORM_TOKEN environment variables. "
-            "Skipping DT MCP enrichment."
-        )
-        return None
-
-    try:
-        from .container.infrastructure.dt_mcp_client import DynatraceMCPClient  # noqa: PLC0415
-    except ImportError as e:
-        logger.warning(f"DT MCP client not available: {e}. Skipping enrichment.")
-        return None
-
-    # Build strategy kwargs
-    strategy_kwargs: dict[str, str] = {}
-    if strategy_name == "llm-query":
-        strategy_kwargs["model"] = model
-
-    try:
-        strategy = create_strategy(strategy_name, **strategy_kwargs)
-    except (ValueError, ImportError) as e:
-        logger.error(f"Failed to create DT strategy '{strategy_name}': {e}")
-        return None
-
-    async def _fetch() -> DtContext | None:
-        mcp_client = DynatraceMCPClient(dt_env, dt_token)
-        try:
-            await mcp_client.connect()
-            logger.info(f"Connected to DT MCP gateway (strategy: {strategy.name})")
-
-            namespace = config.namespace
-            cluster_name = config.kubernetes.context or ""
-            ctx = await strategy.fetch(mcp_client, exploit_type, namespace, cluster_name)
-
-            logger.info(ctx.summary)
-            for q in ctx.queries_executed:
-                logger.debug(f"  DQL: {q[:120]}...")
-
-            return ctx
-        except Exception as e:
-            logger.warning(f"DT MCP enrichment failed: {e}")
-            return None
-        finally:
-            await mcp_client.close()
-
-    return asyncio.run(_fetch())
-
-
-def _revert_applied_resources(
-    k8s: K8sClient, logger: SecurityLogger, namespace: str, dry_run: bool
-) -> None:
-    """Delete all toolkit-labeled resources in a namespace.
-
-    Used to revert ``kimera apply`` operations by removing every resource
-    tagged with ``app.kubernetes.io/managed-by: kimera``.
-    """
-    from .container.infrastructure.resource_applier import (  # noqa: PLC0415
-        TOOLKIT_LABEL,
-        TOOLKIT_LABEL_VALUE,
-    )
-
-    policies = k8s.list_network_policies(namespace)
-    removed = 0
-    for policy in policies:
-        labels = policy.metadata.labels or {}
-        if labels.get(TOOLKIT_LABEL) == TOOLKIT_LABEL_VALUE:
-            if dry_run:
-                logger.info(f"DRY RUN: Would delete NetworkPolicy {policy.metadata.name}")
-            else:
-                k8s.delete_network_policy(policy.metadata.name, namespace)
-            removed += 1
-
-    if removed == 0:
-        logger.info("No toolkit-managed resources found to remove")
-    elif not dry_run:
-        logger.success(f"Removed {removed} toolkit-managed resources")
-
-
-def main() -> None:
-    """Run the main CLI application."""
+def main() -> None:  # noqa: D103
+    from dotenv import load_dotenv
     load_dotenv()
     cli(obj={})
